requirements.txt
faiss-cpu
sentence-transformers
openai
nltk
numpy
sklearn
python-dotenv